ACTUALLY fix the issue with DND
The server and the client very quickly become out of sync, since they
both try to generate UIDs for the assignments. Typing them manually
allows their state to stay in sync. When this data is pulled from the
database or a data store, the issue with generating UIDs will not be an
issue since that happens on the database.